requirements.txt
discord.py[voice]>=2.4.0
PyQt5>=5.15.0
PyNaCl>=1.5.0
